backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
sentence-transformers>=3.0.0
umap-learn>=0.5.6
hdbscan>=0.8.33
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
anthropic>=0.28.0
openai>=1.30.0
weasyprint>=62.0
markdown>=3.6
python-dotenv>=1.0.0
httpx>=0.27.0
